# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eventos API",
        "description": "Descripción de tu API",
        "contact": {
            "email": "roni290690@email.com"
        },
        "license": {
            "name": "Licencia",
            "url": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/eventos/lista-evento": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar Eventos",
                "description": "Reguest",
                "operationId": "listarEvento",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/lista-evento/{idEvento}": {
            "get": {
                "tags": [
                    "Premios"
                ],
                "summary": "Listar Evento por ID",
                "operationId": "showEventoId",
                "parameters": [
                    {
                        "name": "idEvento",
                        "in": "path",
                        "description": "ID del Premio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/create-evento": {
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Crear Evento",
                "operationId": "storeEvento",
                "requestBody": {
                    "description": "Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Eventos"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/update-evento/{idEvento}": {
            "put": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Actualizar Evento",
                "operationId": "updateEvento",
                "parameters": [
                    {
                        "name": "idEvento",
                        "in": "path",
                        "description": "ID del Evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Eventos"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/eliminar-evento/{idEvento}": {
            "delete": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Eliminar Evento",
                "operationId": "destroyEvento",
                "parameters": [
                    {
                        "name": "idEvento",
                        "in": "path",
                        "description": "ID del Evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "datos Elimnados Correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/lista-juego": {
            "get": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Listar Juegos",
                "description": "Reguest",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/lista-juego/{idJuego}": {
            "get": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Listar Juego",
                "operationId": "showJuegoId",
                "parameters": [
                    {
                        "name": "idJuego",
                        "in": "path",
                        "description": "ID del Juego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/create-juego": {
            "post": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Crear Juego",
                "operationId": "storeJuego",
                "requestBody": {
                    "description": "Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Juegos"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/update-juego/{idJuego}": {
            "put": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Actualizar Juego",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "idJuego",
                        "in": "path",
                        "description": "ID del Juego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Juegos"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/eliminar-juego/{idJuego}": {
            "delete": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Eliminar Juego",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "idJuego",
                        "in": "path",
                        "description": "ID del Juego",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "datos Elimnados Correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Eventos": {
                "required": [
                    "nombre",
                    "idUsuario"
                ],
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "idUsuario": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Juegos": {
                "required": [
                    "nombre",
                    "descripcion"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": ""
                    },
                    "descripcion": {
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}